requirements.txt
Pygments==2.12.0
torch
torchvision
pillow
numpy
opencv-python
ipykernel
gitpython
gdown


